Move from s3-benchmark to storage-benchmark
Originally the tool from dvassallo/s3-benchmark was meant to only cover
tests against an S3 endpoint.

Since we also provide file system based tests the name s3-benchmark is
no longer suitable. Instead we're using the more generic name
storage-benchmark.
